Add files via upload
Minesweeper
Minesweeper is a puzzle where there are a number of dangerous squares (mines) spread out across the board. The objective of the player is to determine where the mines lie, and flag them, to warn others. Then to find the safe squares of the board. The game provided allows the player to flag mines, or guess empty squares.

You have been supplied with code that implements the Minesweeper game and also a partially implemented text-based user interface, found in Minesweeper.zip,. The three model classes provided are Assign,

Minesweeper, and Slot. The text-based user interface is UI, and they can be seen in the class diagram. The game is designed to work with supplied puzzle files, these have been included in the Levels folder.

The Tasks
The tasks you are to carry out are as follows:

Provide a full test plan and test log for both the sample code provided (to determine initial functionality), and any further alterations made to the code.
Complete the implementation of the text-based user interface to include the (as yet unimplemented and therefore failed tests) clear (resetting the Minesweeper game), undo (a single user move, then a further move if called again and so on), save (the state of the game to a file) and load (the state of the game from a file) methods. You should only need to make changes to the main user interface class (UI) and the helper class Assign to achieve this. You must use the provided model code, you are not permitted to make your own game from scratch.

design and implement a graphical user interface (GUI) for the Minesweeper model code that allows the user to make moves, undo user moves, clear (reset) the game, save to, and load from a file. The individual slots will need to indicate whether they are unknown, empty, flagged, or an adjacent space. The GUI should at least give feedback on whether the game has been won. The appearance of the interface is up to you. You must use observer-observable to update the GUI as the model changes.
Write a report including design documentation for the full system, including a completed class diagram, descriptions of all classes, and any relevant design decisions made when creating the GUI; and an individual reflection covering the development of the project and evaluating the work.